chore: fix test:coverage errors
The error `No such module "node:inspector"` output during the execution
of the `test:coverage` script was due to a known issue with the
Cloudflare workers vitest integration where it doesn't support using
vitest's v8 coverage support. The fix is to switch to istanbul.

Also includes an update to the reporters configuration that should
hopefully allow the existing dx-team workflow to process the coverage
data from istanbul.